% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-helper.R
\name{.init_helper}
\alias{.init_helper}
\title{Initialize a Helper object}
\usage{
.init_helper(chore = NULL, .chores_chat = get_chores_chat())
}
\arguments{
\item{chore}{The identifier for a helper prompt. By default one
of \link[=cli_helper]{"cli"}, \link[=testthat_helper]{"testthat"} or \link[=roxygen_helper]{"roxygen"},
though custom helpers can be added with \code{\link[=prompt_new]{prompt_new()}}.}

\item{.chores_chat}{An ellmer Chat, e.g.
\code{ellmer::chat_claude()} or \code{ellmer::chat_google_gemini()}.
Defaults to the \code{chores.chat} option, so e.g. set
\code{options(chores.chat = ellmer::chat_claude(model = "claude-haiku-4-5"))}
in your
\code{.Rprofile} to configure chores with ellmer every time you start a new R session.}
}
\value{
A Helper object, which is a subclass of an ellmer chat.
}
\description{
\strong{Users typically should not need to call this function.}
\itemize{
\item Create new helpers that will automatically be registered with this function
with \code{\link[=prompt_new]{prompt_new()}}.
\item The \link[=.init_addin]{chores addin} will initialize needed helpers on-the-fly.
}
}
\examples{
# requires an API key and sets options
\dontrun{
# to create a chat with claude:
.init_helper(.chores_chat = ellmer::chat_claude(model = "claude-haiku-4-5"))

# or with OpenAI's GPT-4.1-mini:
.init_helper(.chores_chat = ellmer::chat_openai(model = "gpt-4.1-mini"))

# or with Google's Gemini 3 Flash:
.init_helper(.chores_chat = ellmer::chat_google_gemini(
  model = "gemini-3-flash-preview",
  api_args = list(
    generationConfig = list(
      thinkingConfig = list(thinkingLevel = "minimal")
    )
  )
))

# to set OpenAI's GPT-4.1-mini as the default model powering chores, for example,
# set the following option (possibly in your .Rprofile, if you'd like
# them to persist across sessions):
options(
  chores.chat = ellmer::chat_openai(model = "gpt-4.1-mini")
)
}
}
